Fix the default trigger
The code expects the @trigger value to be usable for things like:

        mc.$el.add(mc.$trigger).is(":focus")
    
But if the trigger is configured as null (as is the default),
the initTrigger() function will override @$trigger to be
$.noop, which is not usable in the above code line.
    
Therefore, the proper thing to do when trigger is configured to be
null is to set @$trigger to the minical DOM element itself.  That is what
this fix does.